+++ package.json
@@ -24,7 +24,8 @@
     "spanish",
     "german",
     "portuguese",
-    "italian"
+    "italian",
+    "turkish"
   ],
   "author": "Wael TELLAT",
   "license": "MIT",
