+++ test/test_tr.js
@@ -1,0 +1,106 @@
+const assert = require('assert');
+const n2words = require('../index.js');
+
+var testCasesTurkish = [
+  [0, 'sıfır'],
+  [1, 'bir'],
+  [2, 'iki'],
+  [3, 'üç'],
+  [9, "dokuz"],
+  [10, "on"],
+  [11, 'onbir'],
+  [12, 'oniki'],
+  [16, 'onaltı'],
+  [19, 'ondokuz'],
+  [20, 'yirmi'],
+  [21, 'yirmibir'],
+  [26, 'yirmialtı'],
+  [28, 'yirmisekiz'],
+  [30, 'otuz'],
+  [31, 'otuzbir'],
+  [40, 'kırk'],
+  [44, "kırkdört"],
+  [50, 'elli'],
+  [55, 'ellibeş'],
+  [60, 'altmış'],
+  [67, 'altmışyedi'],
+  [70, 'yetmiş'],
+  [79, 'yetmişdokuz'],
+  [89, 'seksendokuz'],
+  [95, 'doksanbeş'],
+  [100, "yüz"],
+  [101, "yüzbir"],
+  [103, "yüzüç"],
+  [110, "yüzon"],
+  [111, "yüzonbir"],
+  [123, "yüzyirmiüç"],
+  [199, 'yüzdoksandokuz'],
+  [203, 'ikiyüzüç'],
+  [287, 'ikiyüzseksenyedi'],
+  [356, 'üçyüzellialtı'],
+  [400, 'dörtyüz'],
+  [434, 'dörtyüzotuzdört'],
+  [578, 'beşyüzyetmişsekiz'],
+  [689, 'altıyüzseksendokuz'],
+  [729, 'yediyüzyirmidokuz'],
+  [894, 'sekizyüzdoksandört'],
+  [999, 'dokuzyüzdoksandokuz'],
+  [1000, 'bin'],
+  [1001, 'binbir'],
+  [1010, "binon"],
+  [1011, "binonbir"],
+  [1097, 'bindoksanyedi'],
+  [1100, "binyüz"],
+  [1104, 'binyüzdört'],
+  [1110, "binyüzon"],
+  [1243, 'binikiyüzkırküç'],
+  [2341, "ikibinüçyüzkırkbir"],
+  [2385, 'ikibinüçyüzseksenbeş'],
+  [3766, 'üçbinyediyüzaltmışaltı'],
+  [4196, 'dörtbinyüzdoksanaltı'],
+  [5846, 'beşbinsekizyüzkırkaltı'],
+  [6459, 'altıbindörtyüzellidokuz'],
+  [7232, 'yedibinikiyüzotuziki'],
+  [8569, 'sekizbinbeşyüzaltmışdokuz'],
+  [9539, 'dokuzbinbeşyüzotuzdokuz'],
+  [10000, "onbin"],
+  [10010, "onbinon"],
+  [10100, "onbinyüz"],
+  [10110, "onbinyüzon"],
+  [11000, "onbirbin"],
+  [35000, "otuzbeşbin"],
+  [116331, "yüzonaltıbinüçyüzotuzbir"],
+  [116330, "yüzonaltıbinüçyüzotuz"],
+  [500000, "beşyüzbin"],
+  [501000, "beşyüzbirbin"],
+  [1000000, 'birmilyon'],
+  [1000001, 'birmilyonbir'],
+  [1000111, "birmilyonyüzonbir"],
+  [4000000, 'dörtmilyon'],
+  [100001000, "yüzmilyonbin"],
+  // [100001001, "yüzmilyonbinbir"],
+  [100010000, "yüzmilyononbin"],
+  [100010001, "yüzmilyononbinbir"],
+  [100011000, "yüzmilyononbirbin"],
+  [100011001, "yüzmilyononbirbinbir"],
+  [101011001, "yüzbirmilyononbirbinbir"],
+  [101011010, "yüzbirmilyononbirbinon"],
+  [111000111, "yüzonbirmilyonyüzonbir"],
+  // [111001111, "yüzonbirmilyonbinyüzonbir"],
+  [111111111, "yüzonbirmilyonyüzonbirbinyüzonbir"],
+  [1101011010, "birmilyaryüzbirmilyononbirbinon"],
+  [1234567890, 'birmilyarikiyüzotuzdörtmilyonbeşyüzaltmışyedibinsekizyüzdoksan'],
+  [101101011010, "yüzbirmilyaryüzbirmilyononbirbinon"],
+  [1000000000001, "birtrilyonbir"],
+  [10000000000000, 'ontrilyon'],
+  [100000000000000, 'yüztrilyon'],
+  [1000000000000000000, 'birkentilyon']
+]
+
+describe('Turkish', function () {
+  it('should convert numbers correctly (Turkish)', function () {
+    for (let i = 0; i < testCasesTurkish.length; i++) {
+      assert.equal(n2words(testCasesTurkish[i][0], { lang: 'tr' }), testCasesTurkish[i][1])
+    }
+  });
+});
